requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crontab==0.7.1
djangorestframework==3.12.2
pika==1.2.0
pytz==2021.1
sqlparse==0.4.1
